+++ knowledge/styles/global.css
@@ -1,3 +1,4 @@
 :root {
   --navbar-height: 76px;
+  --footer-height: 76px;
 }
